+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Added `Symbol.species` to `PaginatedList` class to return `Array` instead of `PaginatedList`
 
 ### Added
 - Download progress bar displayed in the console when connected devices fetch applet content during development